pyproject.toml
[project]
name = "hive-vault"
version = "0.2.0"
description = "Context infrastructure for AI-assisted development — on-demand Obsidian vault access via MCP"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Manuel Lorente" }]
keywords = ["mcp", "claude-code", "obsidian", "vault", "context-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
hive-vault = "hive.vault_server:main"
hive-worker = "hive.worker_server:main"

[project.urls]
Homepage = "https://github.com/mlorentedev/hive"
Repository = "https://github.com/mlorentedev/hive"
Issues = "https://github.com/mlorentedev/hive/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hive"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
